+++ Client/Models/OrderFoodDTOs/OrderDetailCreateDTO.cs
@@ -7,5 +7,5 @@ public class OrderDetailCreateDTO
     public int FoodId { get; set; }
 
     public int Count { get; set; }
-    public double PurchasePrice { get; set; }
+    public decimal PurchasePrice { get; set; }
 }
